python: Don't recompile commands in wasmcoro
Each time wasmcoro.run is called and command instance is Value or
Statement, a new wasm module is generated and compiled. Everytime a new
wasm module is created, a new wasm instance is created too. Wasmtime
limits these instances to 10000 so recompiling existing modules should
be avoided when ever possible.

Fixes: https://github.com/shrine-maiden-heavy-industries/torii-sim-wasm/issues/2